# localhostusak-web/server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "localhost[uşak] Community API",
    "description": "Uşak teknoloji ve tasarım topluluğu web platformu için RESTful API ve SQLite arka plan servisi. Etkinlikler, dinamik etkinlik türleri, kariyer ilanları ve proje vitrini yönetimi.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Yerel Geliştirme Sunucusu"
    }
  ],
  "tags": [
    { "name": "Event Types", "description": "Dinamik Etkinlik Türleri Yönetimi (Cowork, Workshop, Talk vb.)" },
    { "name": "Events", "description": "Topluluk Buluşmaları ve Etkinlikler" },
    { "name": "Careers", "description": "Kariyer Panosu, İş, Staj, Freelance ve Mentorluk İlanları" },
    { "name": "Projects", "description": "Topluluk Projeleri, Vitrin, Ekip Arama ve Açık Kaynak" },
    { "name": "System", "description": "Sistem Sağlık Kontrolü" }
  ],
  "paths": {
    "/api/health": {
      "get": {
        "tags": ["System"],
        "summary": "Sistem sağlık kontrolü",
        "responses": {
          "200": { "description": "Sunucu aktif ve sağlıklı" }
        }
      }
    },
    "/api/event-types": {
      "get": {
        "tags": ["Event Types"],
        "summary": "Tüm etkinlik türlerini listele",
        "description": "Filtre çubuklarında ve etkinlik formlarında kullanılan türleri döner.",
        "responses": {
          "200": { "description": "Etkinlik türleri listesi" }
        }
      }
    },
    "/api/admin/event-types": {
      "post": {
        "tags": ["Event Types"],
        "summary": "Yeni etkinlik türü ekle (Admin)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id", "label"],
                "properties": {
                  "id": { "type": "string", "example": "hackathon" },
                  "label": { "type": "string", "example": "Hackathon" },
                  "icon": { "type": "string", "example": "🏆" },
                  "colorModern": { "type": "string", "example": "#E11D48" },
                  "colorPixel": { "type": "string", "example": "#E11D48" },
                  "sortOrder": { "type": "integer", "example": 4 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Etkinlik türü başarıyla oluşturuldu" },
          "400": { "description": "Eksik parametre" }
        }
      }
    },
    "/api/admin/event-types/{id}": {
      "delete": {
        "tags": ["Event Types"],
        "summary": "Etkinlik türünü sil (Varsayılanlar hariç)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "example": "hackathon" }
        ],
        "responses": {
          "200": { "description": "Etkinlik türü silindi" },
          "400": { "description": "Varsayılan tür silinemez" },
          "404": { "description": "Tür bulunamadı" }
        }
      }
    },
    "/api/events": {
      "get": {
        "tags": ["Events"],
        "summary": "Etkinlikleri listele",
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string" }, "description": "cowork, workshop, talk vb." },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["upcoming", "completed", "cancelled"] } }
        ],
        "responses": {
          "200": { "description": "Etkinlikler listesi" }
        }
      }
    },
    "/api/events/{id}": {
      "get": {
        "tags": ["Events"],
        "summary": "Tek etkinlik detayını getir",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Etkinlik detayı" },
          "404": { "description": "Etkinlik bulunamadı" }
        }
      }
    },
    "/api/admin/events": {
      "post": {
        "tags": ["Events"],
        "summary": "Yeni buluşma / etkinlik oluştur (Admin)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "typeId", "dateStart", "location"],
                "properties": {
                  "title": { "type": "string", "example": "Buluşma #4: AI Agents" },
                  "description": { "type": "string", "example": "Uşak kafe buluşması." },
                  "typeId": { "type": "string", "example": "cowork" },
                  "status": { "type": "string", "example": "upcoming" },
                  "dateStart": { "type": "string", "example": "2026-11-01T14:00:00.000Z" },
                  "location": { "type": "string", "example": "Coff The Story, Uşak" },
                  "capacity": { "type": "integer", "example": 25 },
                  "attendees": { "type": "integer", "example": 0 },
                  "whatsappLink": { "type": "string", "example": "https://chat.whatsapp.com/..." }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Etkinlik oluşturuldu" }
        }
      }
    },
    "/api/admin/events/{id}": {
      "put": {
        "tags": ["Events"],
        "summary": "Etkinliği güncelle (Admin)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Etkinlik güncellendi" }
        }
      },
      "delete": {
        "tags": ["Events"],
        "summary": "Etkinliği sil (Admin)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Etkinlik silindi" }
        }
      }
    },
    "/api/careers": {
      "get": {
        "tags": ["Careers"],
        "summary": "Aktif kariyer ilanlarını listele",
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["job", "internship", "freelance", "mentorship"] } },
          { "name": "work_mode", "in": "query", "schema": { "type": "string", "enum": ["remote", "hybrid", "onsite"] } }
        ],
        "responses": {
          "200": { "description": "Kariyer ilanları listesi" }
        }
      }
    },
    "/api/careers/{id}": {
      "get": {
        "tags": ["Careers"],
        "summary": "Tekil kariyer ilanı detay",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "İlan detayı" }
        }
      }
    },
    "/api/admin/careers": {
      "post": {
        "tags": ["Careers"],
        "summary": "Yeni kariyer ilanı ekle (Admin)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "type", "workMode"],
                "properties": {
                  "title": { "type": "string", "example": "Frontend Developer" },
                  "company": { "type": "string", "example": "Tech Startup" },
                  "type": { "type": "string", "example": "job" },
                  "workMode": { "type": "string", "example": "remote" },
                  "technologies": { "type": "array", "items": { "type": "string" }, "example": ["React", "TypeScript"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "İlan oluşturuldu" }
        }
      }
    },
    "/api/admin/careers/{id}": {
      "delete": {
        "tags": ["Careers"],
        "summary": "Kariyer ilanını sil (Admin)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "İlan silindi" }
        }
      }
    },
    "/api/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "Projeleri listele",
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["showcase", "seeking_team", "opensource"] } }
        ],
        "responses": {
          "200": { "description": "Projeler listesi" }
        }
      }
    },
    "/api/projects/{id}": {
      "get": {
        "tags": ["Projects"],
        "summary": "Proje detayını getir",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Proje detayı" }
        }
      }
    },
    "/api/projects/{id}/like": {
      "patch": {
        "tags": ["Projects"],
        "summary": "Projeyi beğen (+1 like)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Beğeni kaydedildi" }
        }
      }
    },
    "/api/admin/projects": {
      "post": {
        "tags": ["Projects"],
        "summary": "Yeni proje ekle (Admin)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "type"],
                "properties": {
                  "name": { "type": "string", "example": "Uşak Harita Botu" },
                  "description": { "type": "string", "example": "Topluluk projesi." },
                  "type": { "type": "string", "example": "showcase" },
                  "technologies": { "type": "array", "items": { "type": "string" }, "example": ["Python", "FastAPI"] },
                  "owner": { "type": "string", "example": "@dev_member" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Proje oluşturuldu" }
        }
      }
    },
    "/api/admin/projects/{id}": {
      "delete": {
        "tags": ["Projects"],
        "summary": "Projeyi sil (Admin)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Proje silindi" }
        }
      }
    }
  }
}